requirements.txt
streamlit
openai>=1.3.0
pandas
numpy
faiss-cpu
scikit-learn
tqdm
